fix(datasets): enforce deadlines independently of adapter cancellation
